requirements.txt
litellm>=1.40.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
